bookRepoTest complete

+++ src/main/java/com/edu/ulab/app/entity/Book.java
@@ -1,7 +1,6 @@
 package com.edu.ulab.app.entity;
 
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -17,7 +16,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequence")
     @SequenceGenerator(name = "sequence", sequenceName = "sequence", allocationSize = 100)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String title;
